Rozbicie pliku main na poszczególne komponenty. Commit do merge z dev
